requirements.txt
Django==4.2.2
django-bootstrap5==23.3
djangorestframework==3.14.0
environs==8.0.0
Pillow==10.0.0
# psycopg2==2.9.6
python-dotenv==1.0.0

